refactor: use generalized list
