+++ lib/components/List.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { TransactionTable, Header, Table } from '@bpanel/bpanel-ui';
-import Box from './Box';
+import BoxGrid from './BoxGrid';
 
 /*
  * general list component that renders
@@ -37,12 +37,12 @@ class List extends PureComponent {
     else render = <Table colHeaders={headers} tableData={data} />;
 
     return (
-      <Box axis="vertical">
+      <BoxGrid>
         <div>
           <Header type="h5">{text}</Header>
           {render}
         </div>
-      </Box>
+      </BoxGrid>
     );
   }
 }
