--- lib/components/RecentTxList.js
@@ -1,21 +1,0 @@
-import React, { PureComponent } from 'react';
-import { Header } from '@bpanel/bpanel-ui';
-import { TransactionTable } from '@bpanel/bpanel-ui';
-
-class RecentTxList extends PureComponent {
-  constructor(props) {
-    super(props);
-    this.COL_HEADERS = ['Date'];
-  }
-
-  render() {
-    return (
-      <div>
-        <Header type="h5">Recent Tx List</Header>
-        <TransactionTable />
-      </div>
-    );
-  }
-}
-
-export default RecentTxList;
