+++ lib/components/index.js
@@ -1,5 +1,4 @@
-export { default as WalletsList } from './WalletsList';
-export { default as RecentTxList } from './RecentTxList';
+export { default as List } from './List';
 export { default as SendFunds } from './SendFunds';
 export { default as ReceiveFunds } from './ReceiveFunds';
 export { default as NewWallet } from './NewWallet';
