+++ lib/containers/Overview.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Header } from '@bpanel/bpanel-ui';
-import { WalletsList, PendingMultiparty, RecentTxList } from '../components';
+import { List } from '../components';
 
 import { isEqual } from 'lodash';
 
@@ -47,10 +47,9 @@ class Overview extends PureComponent {
     const { walletsList, pendingMultiparty } = this.props;
     return (
       <div>
-        <Header type="h2">Overview</Header>
-        <WalletsList walletsList={walletsList} />
-        <PendingMultiparty pendingMultiparty={pendingMultiparty} />
-        <RecentTxList />
+        <List text="Overview" headers={['Name', 'Balance', 'Watch Only', 'Multi-Sig']} data={walletsList} />
+        <List text="Pending Multiparty Wallets" headers={['Name', 'Date', 'M of N', 'Init']} data={pendingMultiparty} />
+        <List text="Recent Transaction List" type="transaction" />
       </div>
     );
   }
